pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crossplane-watcher-operator"
version = "0.1.0"
description = "Kubernetes operator that watches Crossplane claims and posts status updates on GitHub PRs."
authors = [
    {name = "Cedric Delgehier", email = "cedric.delgehier@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "kubernetes-asyncio>=31.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_provider = "pep621"
tag_format = "v$version"
update_changelog_on_bump = true
major_version_zero = true
version_files = [
    "pyproject.toml:^version",
    "docker/Dockerfile:^ARG APP_VERSION",
]

[dependency-groups]
dev = [
    "commitizen>=3.0.0",
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-v --cov=cwo --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py314"
fix = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["cwo"]
